Components are split into corresponding files.
