+++ src/query-box-app/explorer/documentation/index.tsx
@@ -1,19 +1,8 @@
-import { Badge } from '@/components/ui/badge'
-import { Button } from '@/components/ui/button'
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
-import {
-  Collapsible,
-  CollapsibleContent,
-  CollapsibleTrigger,
-} from '@/components/ui/collapsible'
-import { ChevronDown, ChevronRight, FileText, Hash, Zap } from 'lucide-react'
-import { useMemo, useState } from 'react'
+import { FileText } from 'lucide-react'
+import { Breadcrumb } from './components/breadcrumb'
+import { EmptyState } from './components/empty-state'
+import { TypeContent } from './components/type-content'
 import { useService } from './use-service'
-import {
-  BreadcrumbPathType,
-  CurrentTypeField,
-  DocumentationField,
-} from './utils'
 
 export function Documentation() {
   const {
@@ -58,247 +47,4 @@ export function Documentation() {
       </div>
     </div>
   )
-}
-
-// Breadcrumb Component
-const Breadcrumb = ({
-  path,
-  onNavigate,
-}: {
-  path: BreadcrumbPathType[]
-  onNavigate: (index: number) => void
-}) => (
-  <nav className="flex items-center gap-1 flex-wrap" aria-label="Breadcrumb">
-    {path.map((breadcrumb, idx) => (
-      <div key={breadcrumb.id} className="flex items-center gap-1">
-        <Button
-          variant="ghost"
-          size="sm"
-          className="h-6 px-2 text-xs font-medium hover:bg-accent hover:text-accent-foreground transition-colors rounded"
-          onClick={() => onNavigate(idx)}
-        >
-          {breadcrumb.name}
-        </Button>
-        {idx < path.length - 1 && (
-          <ChevronRight className="h-3 w-3 text-muted-foreground flex-shrink-0" />
-        )}
-      </div>
-    ))}
-  </nav>
-)
-
-// Field Item Component
-const FieldItem = ({
-  field,
-  onNavigate,
-}: {
-  field: DocumentationField
-  onNavigate: (typeName: string) => void
-}) => {
-  return (
-    <div
-      className="group relative flex items-center justify-between py-2.5 px-3 rounded border border-border bg-card transition-all duration-150 hover:bg-accent hover:border-accent-foreground/20 cursor-pointer active:scale-[0.98]"
-      onClick={() => onNavigate(field.type?.name ?? '')}
-      role="button"
-      tabIndex={0}
-      onKeyDown={(e) => {
-        if (e.key === 'Enter' || e.key === ' ') {
-          e.preventDefault()
-          onNavigate(field.type?.name ?? '')
-        }
-      }}
-    >
-      <div className="flex items-center gap-3 min-w-0 flex-1 mr-2">
-        <div className="flex-shrink-0 w-1.5 h-1.5 rounded-full bg-primary/80" />
-        <div className="flex items-center justify-between min-w-0 flex-1 gap-3">
-          <span
-            className="font-medium text-sm text-foreground truncate max-w-[50%]"
-            title={field.name}
-          >
-            {field.name}
-          </span>
-          <Badge
-            variant="secondary"
-            className="text-xs font-mono bg-muted/60 hover:bg-muted/80 transition-colors border-0 px-1.5 py-0.5 flex-shrink-0 max-w-[50%]"
-            title={field.type?.name ?? ''}
-          >
-            <span className="truncate">{field.type?.name ?? ''}</span>
-          </Badge>
-        </div>
-      </div>
-      <ChevronRight className="h-3.5 w-3.5 text-muted-foreground group-hover:text-foreground transition-all duration-150 opacity-0 group-hover:opacity-100 group-hover:translate-x-0.5 flex-shrink-0" />
-    </div>
-  )
-}
-
-// Field List Component
-const FieldList = ({
-  fields,
-  onNavigate,
-}: {
-  fields: DocumentationField[] | undefined
-  onNavigate: (typeName: string) => void
-}) => {
-  if (!fields?.length) return null
-
-  return (
-    <div className="space-y-1.5">
-      {fields.map((field) => (
-        <FieldItem key={field.name} field={field} onNavigate={onNavigate} />
-      ))}
-    </div>
-  )
-}
-
-// Operation Section Component
-const OperationSection = (props: {
-  operation: DocumentationField
-  onNavigate: (typeName: string) => void
-}) => {
-  const { operation, onNavigate } = props
-  const [isOpen, setIsOpen] = useState(true)
-
-  return (
-    <Collapsible open={isOpen} onOpenChange={setIsOpen}>
-      <CollapsibleTrigger asChild>
-        <div className="flex items-center justify-between w-full group cursor-pointer rounded p-1 hover:bg-accent/50 transition-colors">
-          <div className="flex items-center gap-2">
-            <Badge
-              variant="outline"
-              className="font-medium px-2 py-0.5 text-xs bg-background border-border"
-              title={operation.name}
-            >
-              <span className="truncate max-w-[200px]">{operation.name}</span>
-            </Badge>
-          </div>
-          <div className="flex items-center justify-center w-4 h-4 opacity-60 group-hover:opacity-100 transition-opacity">
-            {isOpen ? (
-              <ChevronDown className="h-3 w-3 text-muted-foreground" />
-            ) : (
-              <ChevronRight className="h-3 w-3 text-muted-foreground" />
-            )}
-          </div>
-        </div>
-      </CollapsibleTrigger>
-      <CollapsibleContent>
-        <div key={operation.name} className="space-y-2 mt-1">
-          <div className="pl-3 border-l-2 border-muted">
-            <FieldList fields={operation.subFields} onNavigate={onNavigate} />
-          </div>
-        </div>
-      </CollapsibleContent>
-    </Collapsible>
-  )
-}
-
-// Type Content Component
-const TypeContent = (props: {
-  fields: CurrentTypeField
-  onNavigate: (typeName: string) => void
-}) => {
-  const { fields, onNavigate } = props
-
-  return (
-    <div className="space-y-4">
-      {/* Type Header */}
-      <div className="space-y-2">
-        <div className="flex items-center gap-2">
-          <div className="flex items-center justify-center w-8 h-8 rounded bg-primary/10">
-            <Hash className="h-4 w-4 text-primary" />
-          </div>
-          <h1 className="text-xl font-bold text-foreground">{fields.name}</h1>
-        </div>
-        {fields.description && (
-          <p className="text-sm text-muted-foreground leading-relaxed pl-10">
-            <ParsedDescription text={fields.description} />
-          </p>
-        )}
-      </div>
-
-      {/* Operations Section */}
-      {fields.operations?.some((op) => op.subFields?.length) && (
-        <Card className="shadow-sm gap-2">
-          <CardHeader className="pb-2">
-            <CardTitle className="text-lg font-semibold flex items-center gap-2">
-              <div className="flex items-center justify-center w-6 h-6 rounded bg-primary/10">
-                <Zap className="h-3.5 w-3.5 text-primary" />
-              </div>
-              Operations
-            </CardTitle>
-          </CardHeader>
-          <CardContent className="space-y-4">
-            {fields.operations?.map((operation) =>
-              operation.subFields?.length ? (
-                <OperationSection
-                  key={operation.name}
-                  operation={operation}
-                  onNavigate={onNavigate}
-                />
-              ) : null
-            )}
-          </CardContent>
-        </Card>
-      )}
-
-      {/* Fields Section */}
-      {fields.subFields?.length ? (
-        <Card className="shadow-sm gap-2">
-          <CardHeader className="pb-2">
-            <CardTitle className="text-lg font-semibold flex items-center gap-2">
-              <div className="flex items-center justify-center w-6 h-6 rounded bg-primary/10">
-                <FileText className="h-3.5 w-3.5 text-primary" />
-              </div>
-              Fields
-            </CardTitle>
-          </CardHeader>
-          <CardContent>
-            <FieldList fields={fields.subFields} onNavigate={onNavigate} />
-          </CardContent>
-        </Card>
-      ) : null}
-    </div>
-  )
-}
-
-// Empty State Component
-const EmptyState = ({ typeName }: { typeName: string }) => (
-  <div className="flex flex-col items-center justify-center py-12 text-center">
-    <div className="rounded-full bg-muted/50 p-4 mb-4">
-      <FileText className="h-8 w-8 text-muted-foreground" />
-    </div>
-    <div className="space-y-2 max-w-sm">
-      <h3 className="font-semibold text-lg text-foreground">
-        Unsupported Type
-      </h3>
-      <p className="text-sm text-muted-foreground leading-relaxed">
-        This type is not currently supported for documentation viewing.
-      </p>
-      <Badge
-        variant="secondary"
-        className="font-mono px-3 py-1 bg-muted/60 border-0 text-xs"
-      >
-        {typeName}
-      </Badge>
-    </div>
-  </div>
-)
-
-const ParsedDescription = (props: { text: string }) => {
-  const { text } = props
-  const parts = useMemo(() => text.split(/(`[^`]+`)/g), [text])
-
-  return parts.map((part, index) => {
-    if (part.startsWith('`') && part.endsWith('`')) {
-      const code = part.slice(1, -1)
-      return (
-        <code
-          key={index}
-          className="relative rounded bg-muted px-[0.3rem] py-[0.2rem] font-mono text-sm font-semibold"
-        >
-          {code}
-        </code>
-      )
-    }
-    return <span key={index}>{part}</span>
-  })
 }
